Clean up works * Deprecated `ion_nature` * Migrated objects construction from `from_value()` to their constructors in C extension * Added IonPytimestamp and IonPyBytes * Replaced all the original IonPyObject and factory methods with the new IonPyObjects. * Worked on the IonPyTimestamp signature to make it compatible with the amazon.core.Timestamp * Updated the unit tests to accommodate the changes.
