+++ amazon/ion/ioncmodule.c
@@ -38,27 +38,16 @@ static PyObject* _decimal_constructor;
 static PyObject* _py_timestamp_constructor;
 static PyObject* _simpletypes_module;
 static PyObject* _ionpynull_cls;
-static PyObject* _ionpynull_fromvalue;
 static PyObject* _ionpybool_cls;
-static PyObject* _ionpybool_fromvalue;
 static PyObject* _ionpyint_cls;
-static PyObject* _ionpyint_fromvalue;
 static PyObject* _ionpyfloat_cls;
-static PyObject* _ionpyfloat_fromvalue;
 static PyObject* _ionpydecimal_cls;
-static PyObject* _ionpydecimal_fromvalue;
 static PyObject* _ionpytimestamp_cls;
-static PyObject* _ionpytimestamp_fromvalue;
 static PyObject* _ionpytext_cls;
-static PyObject* _ionpytext_fromvalue;
 static PyObject* _ionpysymbol_cls;
-static PyObject* _ionpysymbol_fromvalue;
 static PyObject* _ionpybytes_cls;
-static PyObject* _ionpybytes_fromvalue;
 static PyObject* _ionpylist_cls;
-static PyObject* _ionpylist_fromvalue;
 static PyObject* _ionpydict_cls;
-static PyObject* _ionpydict_fromvalue;
 static PyObject* _ion_core_module;
 static PyObject* _py_ion_type;
 static PyObject* py_ion_type_table[14];
@@ -1115,7 +1104,7 @@ iERR ionc_read_value(hREADER hreader, ION_TYPE t, PyObject* container, BOOL in_s
             ion_type = ION_TYPE_INT(null_type);
             py_value = Py_None; // INCREFs and returns Python None.
             emit_bare_values = emit_bare_values && (ion_type == tid_NULL_INT);
-            ion_nature_constructor = _ionpynull_fromvalue;
+            ion_nature_constructor = _ionpynull_cls;
             break;
         }
         case tid_BOOL_INT:
@@ -1123,7 +1112,7 @@ iERR ionc_read_value(hREADER hreader, ION_TYPE t, PyObject* container, BOOL in_s
             BOOL bool_value;
             IONCHECK(ion_reader_read_bool(hreader, &bool_value));
             py_value = PyBool_FromLong(bool_value);
-            ion_nature_constructor = _ionpybool_fromvalue;
+            ion_nature_constructor = _ionpybool_cls;
             break;
         }
         case tid_INT_INT:
@@ -1146,7 +1135,7 @@ iERR ionc_read_value(hREADER hreader, ION_TYPE t, PyObject* container, BOOL in_s
             py_value = PyLong_FromString(ion_int_str, NULL, 10);
             PyMem_Free(ion_int_str);
 
-            ion_nature_constructor = _ionpyint_fromvalue;
+            ion_nature_constructor = _ionpyint_cls;
             break;
         }
         case tid_FLOAT_INT:
@@ -1154,7 +1143,7 @@ iERR ionc_read_value(hREADER hreader, ION_TYPE t, PyObject* container, BOOL in_s
             double double_value;
             IONCHECK(ion_reader_read_double(hreader, &double_value));
             py_value = Py_BuildValue("d", double_value);
-            ion_nature_constructor = _ionpyfloat_fromvalue;
+            ion_nature_constructor = _ionpyfloat_cls;
             break;
         }
         case tid_DECIMAL_INT:
@@ -1209,13 +1198,13 @@ iERR ionc_read_value(hREADER hreader, ION_TYPE t, PyObject* container, BOOL in_s
                 }
             }
 
-            ion_nature_constructor = _ionpydecimal_fromvalue;
+            ion_nature_constructor = _ionpydecimal_cls;
             break;
         }
         case tid_TIMESTAMP_INT:
         {
             IONCHECK(ionc_read_timestamp(hreader, &py_value));
-            ion_nature_constructor = _ionpytimestamp_fromvalue;
+            ion_nature_constructor = _ionpytimestamp_cls;
             break;
         }
         case tid_SYMBOL_INT:
@@ -1223,7 +1212,7 @@ iERR ionc_read_value(hREADER hreader, ION_TYPE t, PyObject* container, BOOL in_s
             emit_bare_values = FALSE; // Symbol values must always be emitted as IonNature because of ambiguity with string.
             ION_STRING string_value;
             IONCHECK(ion_reader_read_string(hreader, &string_value));
-            ion_nature_constructor = _ionpysymbol_fromvalue;
+            ion_nature_constructor = _ionpysymbol_cls;
             py_value = ion_string_to_py_symboltoken(&string_value);
             break;
         }
@@ -1232,7 +1221,7 @@ iERR ionc_read_value(hREADER hreader, ION_TYPE t, PyObject* container, BOOL in_s
             ION_STRING string_value;
             IONCHECK(ion_reader_read_string(hreader, &string_value));
             py_value = ion_build_py_string(&string_value);
-            ion_nature_constructor = _ionpytext_fromvalue;
+            ion_nature_constructor = _ionpytext_cls;
             break;
         }
         case tid_CLOB_INT:
@@ -1264,12 +1253,12 @@ iERR ionc_read_value(hREADER hreader, ION_TYPE t, PyObject* container, BOOL in_s
             if (length) {
                 PyMem_Free(buf);
             }
-            ion_nature_constructor = _ionpybytes_fromvalue;
+            ion_nature_constructor = _ionpybytes_cls;
             break;
         }
         case tid_STRUCT_INT:
         {
-            ion_nature_constructor = _ionpydict_fromvalue;
+            ion_nature_constructor = _ionpydict_cls;
             //Init a IonPyDict
             PyObject* new_dict = PyDict_New();
             py_value = PyObject_CallFunctionObjArgs(
@@ -1294,7 +1283,7 @@ iERR ionc_read_value(hREADER hreader, ION_TYPE t, PyObject* container, BOOL in_s
         {
             py_value = PyList_New(0);
             IONCHECK(ionc_read_into_container(hreader, py_value, /*is_struct=*/FALSE, emit_bare_values));
-            ion_nature_constructor = _ionpylist_fromvalue;
+            ion_nature_constructor = _ionpylist_cls;
             break;
         }
         case tid_DATAGRAM_INT:
@@ -1556,39 +1545,17 @@ PyObject* ionc_init_module(void) {
     _simpletypes_module         = PyImport_ImportModule("amazon.ion.simple_types");
 
     _ionpynull_cls              = PyObject_GetAttrString(_simpletypes_module, "IonPyNull");
-    _ionpynull_fromvalue        = PyObject_GetAttrString(_ionpynull_cls, "from_value");
     _ionpybool_cls              = PyObject_GetAttrString(_simpletypes_module, "IonPyBool");
-    _ionpybool_fromvalue        = PyObject_GetAttrString(_ionpybool_cls, "from_value");
     _ionpyint_cls               = PyObject_GetAttrString(_simpletypes_module, "IonPyInt");
-    _ionpyint_fromvalue         = PyObject_GetAttrString(_ionpyint_cls, "from_value");
     _ionpyfloat_cls             = PyObject_GetAttrString(_simpletypes_module, "IonPyFloat");
-    _ionpyfloat_fromvalue       = PyObject_GetAttrString(_ionpyfloat_cls, "from_value");
     _ionpydecimal_cls           = PyObject_GetAttrString(_simpletypes_module, "IonPyDecimal");
-    _ionpydecimal_fromvalue     = PyObject_GetAttrString(_ionpydecimal_cls, "from_value");
     _ionpytimestamp_cls         = PyObject_GetAttrString(_simpletypes_module, "IonPyTimestamp");
-    _ionpytimestamp_fromvalue   = PyObject_GetAttrString(_ionpytimestamp_cls, "from_value");
     _ionpybytes_cls             = PyObject_GetAttrString(_simpletypes_module, "IonPyBytes");
-    _ionpybytes_fromvalue       = PyObject_GetAttrString(_ionpybytes_cls, "from_value");
     _ionpytext_cls              = PyObject_GetAttrString(_simpletypes_module, "IonPyText");
-    _ionpytext_fromvalue        = PyObject_GetAttrString(_ionpytext_cls, "from_value");
     _ionpysymbol_cls            = PyObject_GetAttrString(_simpletypes_module, "IonPySymbol");
-    _ionpysymbol_fromvalue      = PyObject_GetAttrString(_ionpysymbol_cls, "from_value");
     _ionpylist_cls              = PyObject_GetAttrString(_simpletypes_module, "IonPyList");
-    _ionpylist_fromvalue        = PyObject_GetAttrString(_ionpylist_cls, "from_value");
     _ionpydict_cls              = PyObject_GetAttrString(_simpletypes_module, "IonPyDict");
-    _ionpydict_fromvalue        = PyObject_GetAttrString(_ionpydict_cls, "from_value");
 
-    // A temporary workaround to replace from_value() with class constructors
-    // TODO deprecate from_value later
-    _ionpyint_fromvalue = _ionpyint_cls;
-    _ionpynull_fromvalue = _ionpynull_cls;
-    _ionpyfloat_fromvalue = _ionpyfloat_cls;
-    _ionpydecimal_fromvalue = _ionpydecimal_cls;
-    _ionpytext_fromvalue = _ionpytext_cls;
-    _ionpylist_fromvalue = _ionpylist_cls;
-    _ionpydict_fromvalue = _ionpydict_cls;
-    _ionpybool_fromvalue = _ionpybool_cls;
-    _ionpysymbol_fromvalue = _ionpysymbol_cls;
 
     _ion_core_module            = PyImport_ImportModule("amazon.ion.core");
     _py_timestamp_precision     = PyObject_GetAttrString(_ion_core_module, "TimestampPrecision");
