+++ amazon/ion/simple_types.py
@@ -20,9 +20,6 @@ particulars of the Ion data model.
 
 from decimal import Decimal
 
-from amazon.ion.ion_py_objects import IonPyNull, IonPyDecimal, IonPyInt, IonPyFloat, IonPyText, \
-    IonPyList, IonPyDict, IonPyBool, IonPySymbol
-
 # in Python 3.10, abstract collections have moved into their own module
 # for compatibility with 3.10+, first try imports from the new location
 # if that fails, try from the pre-3.10 location
@@ -31,36 +28,26 @@ try:
 except:
     from collections import MutableMapping
 
+from collections import OrderedDict
+from decimal import Decimal
+
+from amazon.ion.core import IonType, IonEvent, Timestamp, TIMESTAMP_FRACTIONAL_SECONDS_FIELD, TIMESTAMP_PRECISION_FIELD, \
+    TimestampPrecision
 from amazon.ion.symbols import SymbolToken
-from .core import TIMESTAMP_PRECISION_FIELD
-from .core import Multimap, Timestamp, IonEvent, IonType, TIMESTAMP_FRACTION_PRECISION_FIELD, TimestampPrecision, \
-    MICROSECOND_PRECISION, TIMESTAMP_FRACTIONAL_SECONDS_FIELD
 
 
-class _IonNature(object):
-    """Mix-in for Ion related properties.
+class IonPyNull(object):
+    __name__ = 'IonPyNull'
+    __qualname__ = 'IonPyNull'
 
-    Attributes:
-        ion_event (Optional[IonEvent]): The event, if any associated with the value.
-        ion_type (IonType): The Ion type for the value.
-        ion_annotations (Sequence[unicode]): The annotations associated with the value.
+    def __init__(self, ion_type=IonType.NULL, value=None, annotations=()):
+        self.ion_type = ion_type
+        self.ion_annotations = annotations
 
-    Notes:
-        There is no ``field_name`` attribute as that is generally modeled as a property of the
-        container.
+    def __bool__(self):
+        return False
 
-        The ``ion_event`` field is only provided if the value was derived from a low-level event.
-        User constructed values will generally not set this field.
-    """
-    def __init__(self, *args, **kwargs):
-        self.ion_type = None
-        self.ion_annotations = ()
-
-    def _copy(self):
-        """Copies this instance. Its IonEvent (if any) is not preserved.
-
-        Keeping this protected until/unless we decide there's use for it publicly.
-        """
+    def __copy__(self):
         args, kwargs = self._to_constructor_args(self)
         value = self.__class__(*args, **kwargs)
         value.ion_type = self.ion_type
@@ -69,21 +56,14 @@ class _IonNature(object):
 
     @staticmethod
     def _to_constructor_args(value):
-        return (value, ), {}
+        return (None, value,), {}
 
     @classmethod
     def from_event(cls, ion_event):
-        """Constructs the given native extension from the properties of an event.
-
-        Args:
-            ion_event (IonEvent): The event to construct the native value from.
-        """
         if ion_event.value is not None:
             args, kwargs = cls._to_constructor_args(ion_event.value)
         else:
-            # if value is None (i.e. this is a container event), args must be empty or initialization of the
-            # underlying container will fail.
-            args, kwargs = (), {}
+            args, kwargs = (None, None, ()), {}
         value = cls(*args, **kwargs)
         value.ion_type = ion_event.ion_type
         value.ion_annotations = ion_event.annotations
@@ -91,13 +71,6 @@ class _IonNature(object):
 
     @classmethod
     def from_value(cls, ion_type, value, annotations=()):
-        """Constructs a value as a copy with an associated Ion type and annotations.
-
-        Args:
-            ion_type (IonType): The associated Ion type.
-            value (Any): The value to construct from, generally of type ``cls``.
-            annotations (Sequence[unicode]):  The sequence Unicode strings decorating this value.
-        """
         if value is None:
             value = IonPyNull()
         else:
@@ -108,22 +81,8 @@ class _IonNature(object):
         return value
 
     def to_event(self, event_type, field_name=None, in_struct=False, depth=None):
-        """Constructs an IonEvent from this _IonNature value.
-
-        Args:
-            event_type (IonEventType): The type of the resulting event.
-            field_name (Optional[text]): The field name associated with this value, if any.  When ``None``
-                is specified and ``in_struct`` is ``True``, the returned event's ``field_name`` will
-                represent symbol zero (a ``SymbolToken`` with text=None and sid=0).
-            in_struct (Optional[True|False]): When ``True``, indicates the returned event ``field_name``
-                will be populated.  When ``False``, ``field_name`` will be ``None``.
-            depth (Optional[int]): The depth of this value.
-
-        Returns:
-            An IonEvent with the properties from this value.
-        """
         value = self
-        if isinstance(self, IonPyNull) or self.ion_type.is_container:
+        if isinstance(self, IonPyNull):
             value = None
 
         if in_struct:
@@ -133,51 +92,390 @@ class _IonNature(object):
             field_name = None
 
         return IonEvent(event_type, ion_type=self.ion_type, value=value, field_name=field_name,
-                                  annotations=self.ion_annotations, depth=depth)
+                        annotations=self.ion_annotations, depth=depth)
 
 
-def _ion_type_for(name, base_cls, ion_type=None):
-    class IonPyValueType(base_cls, _IonNature):
-        def __init__(self, *args, **kwargs):
-            super(IonPyValueType, self).__init__(*args, **kwargs)
-            self.ion_type = ion_type
+class IonPyDecimal(Decimal):
+    __name__ = 'IonPyDecimal'
+    __qualname__ = 'IonPyDecimal'
+    ion_type = IonType.DECIMAL
 
-    IonPyValueType.__name__ = name
-    IonPyValueType.__qualname__ = name
-    return IonPyValueType
+    def __new__(cls, ion_type=IonType.DECIMAL, value=None, annotations=()):
+        v = super().__new__(cls, value)
+        v.ion_annotations = annotations
+        return v
 
-
-# All of our ion value types inherit from their python value types, for good or
-# ill. Python's builtin `bool` cannot be sub-classed and is itself a sub-class
-# of int so we just sub-class int directly for our Ion Boolean.
-# Considering that True == 1 and False == 0 this works as expected.
-IonPyBool_original = _ion_type_for('IonPyBool', int, IonType.BOOL)
-
-# We override __repr__ so Booleans show up as True|False and not 1|0.
-# The representation is consistent with other primitives in that it returns a
-# string repr of the value but not annotations.
-IonPyBool_original.__repr__ = lambda self: str(bool(self))
-
-
-class IonPySymbol_original(SymbolToken, _IonNature):
-    def __init__(self, *args, **kwargs):
-        super(IonPySymbol_original, self).__init__(*args, **kwargs)
-        self.ion_type = IonType.SYMBOL
+    def __copy__(self):
+        args, kwargs = self._to_constructor_args(self)
+        value = self.__class__(*args, **kwargs)
+        value.ion_type = self.ion_type
+        value.ion_annotations = self.ion_annotations
+        return value
 
     @staticmethod
-    def _to_constructor_args(st):
-        try:
-            args = (st.text, st.sid, st.location)
-        except AttributeError:
-            args = (st, None, None)
-        kwargs = {}
-        return args, kwargs
+    def _to_constructor_args(value):
+        return (None, value,), {}
+
+    @classmethod
+    def from_event(cls, ion_event):
+        if ion_event.value is not None:
+            args, kwargs = cls._to_constructor_args(ion_event.value)
+        else:
+            args, kwargs = (None, None, ()), {}
+        value = cls(*args, **kwargs)
+        value.ion_type = ion_event.ion_type
+        value.ion_annotations = ion_event.annotations
+        return value
+
+    @classmethod
+    def from_value(cls, ion_type, value, annotations=()):
+        if value is None:
+            value = IonPyNull()
+        else:
+            args, kwargs = cls._to_constructor_args(value)
+            value = cls(*args, **kwargs)
+        value.ion_type = ion_type
+        value.ion_annotations = annotations
+        return value
+
+    def to_event(self, event_type, field_name=None, in_struct=False, depth=None):
+        value = self
+        if isinstance(self, IonPyNull):
+            value = None
+
+        if in_struct:
+            if not isinstance(field_name, SymbolToken):
+                field_name = SymbolToken(field_name, 0 if field_name is None else None)
+        else:
+            field_name = None
+
+        return IonEvent(event_type, ion_type=self.ion_type, value=value, field_name=field_name,
+                        annotations=self.ion_annotations, depth=depth)
 
 
-class IonPyTimestamp(Timestamp, _IonNature):
-    def __init__(self, *args, **kwargs):
-        super(IonPyTimestamp, self).__init__(*args, **kwargs)
-        self.ion_type = IonType.TIMESTAMP
+class IonPyBytes(bytes):
+    __name__ = 'IonPyBytes'
+    __qualname__ = 'IonPyBytes'
+
+    def __new__(cls, ion_type=IonType.BLOB, value=None, annotations=()):
+        v = super().__new__(cls, value)
+        v.ion_annotations = annotations
+        v.ion_type = ion_type
+        return v
+
+    def __copy__(self):
+        args, kwargs = self._to_constructor_args(self)
+        value = self.__class__(*args, **kwargs)
+        value.ion_type = self.ion_type
+        value.ion_annotations = self.ion_annotations
+        return value
+
+    @staticmethod
+    def _to_constructor_args(value):
+        return (None, value,), {}
+
+    @classmethod
+    def from_event(cls, ion_event):
+        if ion_event.value is not None:
+            args, kwargs = cls._to_constructor_args(ion_event.value)
+        else:
+            args, kwargs = (None, None, ()), {}
+        value = cls(*args, **kwargs)
+        value.ion_type = ion_event.ion_type
+        value.ion_annotations = ion_event.annotations
+        return value
+
+    @classmethod
+    def from_value(cls, ion_type, value, annotations=()):
+        if value is None:
+            value = IonPyNull()
+        else:
+            args, kwargs = cls._to_constructor_args(value)
+            value = cls(*args, **kwargs)
+        value.ion_type = ion_type
+        value.ion_annotations = annotations
+        return value
+
+    def to_event(self, event_type, field_name=None, in_struct=False, depth=None):
+        value = self
+        if isinstance(self, IonPyNull):
+            value = None
+
+        if in_struct:
+            if not isinstance(field_name, SymbolToken):
+                field_name = SymbolToken(field_name, 0 if field_name is None else None)
+        else:
+            field_name = None
+
+        return IonEvent(event_type, ion_type=self.ion_type, value=value, field_name=field_name,
+                        annotations=self.ion_annotations, depth=depth)
+
+
+class IonPyInt(int):
+    __name__ = 'IonPyInt'
+    __qualname__ = 'IonPyInt'
+    ion_type = IonType.INT
+
+    def __new__(cls, ion_type=IonType.INT, value=None, annotations=()):
+        v = super().__new__(cls, value)
+        v.ion_annotations = annotations
+        return v
+
+    def __copy__(self):
+        args, kwargs = self._to_constructor_args(self)
+        value = self.__class__(*args, **kwargs)
+        value.ion_type = self.ion_type
+        value.ion_annotations = self.ion_annotations
+        return value
+
+    @staticmethod
+    def _to_constructor_args(value):
+        return (None, value,), {}
+
+    @classmethod
+    def from_event(cls, ion_event):
+        if ion_event.value is not None:
+            args, kwargs = cls._to_constructor_args(ion_event.value)
+        else:
+            args, kwargs = (None, None, ()), {}
+        value = cls(*args, **kwargs)
+        value.ion_type = ion_event.ion_type
+        value.ion_annotations = ion_event.annotations
+        return value
+
+    @classmethod
+    def from_value(cls, ion_type, value, annotations=()):
+        if value is None:
+            value = IonPyNull()
+        else:
+            args, kwargs = cls._to_constructor_args(value)
+            value = cls(*args, **kwargs)
+        value.ion_type = ion_type
+        value.ion_annotations = annotations
+        return value
+
+    def to_event(self, event_type, field_name=None, in_struct=False, depth=None):
+        value = self
+        if isinstance(self, IonPyNull):
+            value = None
+
+        if in_struct:
+            if not isinstance(field_name, SymbolToken):
+                field_name = SymbolToken(field_name, 0 if field_name is None else None)
+        else:
+            field_name = None
+
+        return IonEvent(event_type, ion_type=self.ion_type, value=value, field_name=field_name,
+                        annotations=self.ion_annotations, depth=depth)
+
+
+class IonPyBool(int):
+    __name__ = 'IonPyBool'
+    __qualname__ = 'IonPyBool'
+    ion_type = IonType.BOOL
+
+    def __repr__(self):
+        return str(bool(self))
+
+    def __new__(cls, ion_type=IonType.BOOL, value=None, annotations=()):
+        v = super().__new__(cls, value)
+        v.ion_annotations = annotations
+        return v
+
+    def __copy__(self):
+        args, kwargs = self._to_constructor_args(self)
+        value = self.__class__(*args, **kwargs)
+        value.ion_type = self.ion_type
+        value.ion_annotations = self.ion_annotations
+        return value
+
+    @staticmethod
+    def _to_constructor_args(value):
+        return (None, value,), {}
+
+    @classmethod
+    def from_event(cls, ion_event):
+        if ion_event.value is not None:
+            args, kwargs = cls._to_constructor_args(ion_event.value)
+        else:
+            args, kwargs = (None, None, ()), {}
+        value = cls(*args, **kwargs)
+        value.ion_type = ion_event.ion_type
+        value.ion_annotations = ion_event.annotations
+        return value
+
+    @classmethod
+    def from_value(cls, ion_type, value, annotations=()):
+        if value is None:
+            value = IonPyNull()
+        else:
+            args, kwargs = cls._to_constructor_args(value)
+            value = cls(*args, **kwargs)
+        value.ion_type = ion_type
+        value.ion_annotations = annotations
+        return value
+
+    def to_event(self, event_type, field_name=None, in_struct=False, depth=None):
+        value = self
+        if isinstance(self, IonPyNull):
+            value = None
+
+        if in_struct:
+            if not isinstance(field_name, SymbolToken):
+                field_name = SymbolToken(field_name, 0 if field_name is None else None)
+        else:
+            field_name = None
+
+        return IonEvent(event_type, ion_type=self.ion_type, value=value, field_name=field_name,
+                        annotations=self.ion_annotations, depth=depth)
+
+
+class IonPyFloat(float):
+    __name__ = 'IonPyFloat'
+    __qualname__ = 'IonPyFloat'
+    ion_type = IonType.FLOAT
+
+    def __new__(cls, ion_type=IonType.FLOAT, value=None, annotations=()):
+        v = super().__new__(cls, value)
+        v.ion_annotations = annotations
+        return v
+
+    def __copy__(self):
+        args, kwargs = self._to_constructor_args(self)
+        value = self.__class__(*args, **kwargs)
+        value.ion_type = self.ion_type
+        value.ion_annotations = self.ion_annotations
+        return value
+
+    @staticmethod
+    def _to_constructor_args(value):
+        return (None, value,), {}
+
+    @classmethod
+    def from_event(cls, ion_event):
+        if ion_event.value is not None:
+            args, kwargs = cls._to_constructor_args(ion_event.value)
+        else:
+            args, kwargs = (None, None, ()), {}
+        value = cls(*args, **kwargs)
+        value.ion_type = ion_event.ion_type
+        value.ion_annotations = ion_event.annotations
+        return value
+
+    @classmethod
+    def from_value(cls, ion_type, value, annotations=()):
+        if value is None:
+            value = IonPyNull()
+        else:
+            args, kwargs = cls._to_constructor_args(value)
+            value = cls(*args, **kwargs)
+        value.ion_type = ion_type
+        value.ion_annotations = annotations
+        return value
+
+    def to_event(self, event_type, field_name=None, in_struct=False, depth=None):
+        value = self
+        if isinstance(self, IonPyNull):
+            value = None
+
+        if in_struct:
+            if not isinstance(field_name, SymbolToken):
+                field_name = SymbolToken(field_name, 0 if field_name is None else None)
+        else:
+            field_name = None
+
+        return IonEvent(event_type, ion_type=self.ion_type, value=value, field_name=field_name,
+                        annotations=self.ion_annotations, depth=depth)
+
+
+class IonPyText(str):
+    __name__ = 'IonPyText'
+    __qualname__ = 'IonPyText'
+    ion_type = IonType.STRING
+
+    def __new__(cls, ion_type=IonType.STRING, value=None, annotations=()):
+        v = super().__new__(cls, value)
+        v.ion_annotations = annotations
+        return v
+
+    def __copy__(self):
+        args, kwargs = self._to_constructor_args(self)
+        value = self.__class__(*args, **kwargs)
+        value.ion_type = self.ion_type
+        value.ion_annotations = self.ion_annotations
+        return value
+
+    @staticmethod
+    def _to_constructor_args(value):
+        return (None, value,), {}
+
+    @classmethod
+    def from_event(cls, ion_event):
+        if ion_event.value is not None:
+            args, kwargs = cls._to_constructor_args(ion_event.value)
+        else:
+            args, kwargs = (None, None, ()), {}
+        value = cls(*args, **kwargs)
+        value.ion_type = ion_event.ion_type
+        value.ion_annotations = ion_event.annotations
+        return value
+
+    @classmethod
+    def from_value(cls, ion_type, value, annotations=()):
+        if value is None:
+            value = IonPyNull()
+        else:
+            args, kwargs = cls._to_constructor_args(value)
+            value = cls(*args, **kwargs)
+        value.ion_type = ion_type
+        value.ion_annotations = annotations
+        return value
+
+    def to_event(self, event_type, field_name=None, in_struct=False, depth=None):
+        value = self
+        if isinstance(self, IonPyNull):
+            value = None
+
+        if in_struct:
+            if not isinstance(field_name, SymbolToken):
+                field_name = SymbolToken(field_name, 0 if field_name is None else None)
+        else:
+            field_name = None
+
+        return IonEvent(event_type, ion_type=self.ion_type, value=value, field_name=field_name,
+                        annotations=self.ion_annotations, depth=depth)
+
+
+class IonPyTimestamp(Timestamp):
+    __name__ = 'IonPyTimestamp'
+    __qualname__ = 'IonPyTimestamp'
+    ion_type = IonType.TIMESTAMP
+
+    def __new__(cls, *args, **kwargs):
+        # The from_value like signature
+        if isinstance(args[0], IonType):
+            value = args[1]
+            annotations = ()
+            if len(args) > 2 and args[2] is not None:
+                annotations = args[2]
+            if value is not None:
+                args_new, kwargs = cls._to_constructor_args(value)
+            else:
+                args_new, kwargs = (None, None, ()), {}
+            v = super().__new__(cls, *args_new, **kwargs)
+            v.ion_type = args[0]
+            v.ion_annotations = annotations
+        # Regular timestamp constructor
+        else:
+            v = super().__new__(cls, *args, **kwargs)
+        return v
+
+    def __copy__(self):
+        args, kwargs = self._to_constructor_args(self)
+        value = self.__class__(*args, **kwargs)
+        value.ion_type = self.ion_type
+        value.ion_annotations = self.ion_annotations
+        return value
 
     @staticmethod
     def _to_constructor_args(ts):
@@ -191,61 +489,301 @@ class IonPyTimestamp(Timestamp, _IonNature):
             kwargs = {TIMESTAMP_PRECISION_FIELD: TimestampPrecision.SECOND}
         return args, kwargs
 
+    @classmethod
+    def from_event(cls, ion_event):
+        if ion_event.value is not None:
+            args, kwargs = cls._to_constructor_args(ion_event.value)
+        else:
+            args, kwargs = (None, None, ()), {}
+        value = cls(*args, **kwargs)
+        value.ion_type = ion_event.ion_type
+        value.ion_annotations = ion_event.annotations
+        return value
 
-class IonPyNull_original(_IonNature):
-    """Representation of ``null``.
+    @classmethod
+    def from_value(cls, ion_type, value, annotations=()):
+        if value is None:
+            value = IonPyNull()
+        else:
+            args, kwargs = cls._to_constructor_args(value)
+            value = cls(*args, **kwargs)
+        value.ion_type = ion_type
+        value.ion_annotations = annotations
+        return value
 
-    Notes:
-        ``None`` is a singleton and cannot be sub-classed, so we have our
-         own value type for it.  The function ``is_null`` is the best way
-         to test for ``null``-ness or ``None``-ness.
-    """
+    def to_event(self, event_type, field_name=None, in_struct=False, depth=None):
+        value = self
+        if isinstance(self, IonPyNull):
+            value = None
 
-    def __init__(self, *args, **kwargs):
-        super(IonPyNull_original, self).__init__(*args, **kwargs)
+        if in_struct:
+            if not isinstance(field_name, SymbolToken):
+                field_name = SymbolToken(field_name, 0 if field_name is None else None)
+        else:
+            field_name = None
 
-    def __nonzero__(self):
-        return False
+        return IonEvent(event_type, ion_type=self.ion_type, value=value, field_name=field_name,
+                        annotations=self.ion_annotations, depth=depth)
 
-    def __bool__(self):
-        return False
+
+class IonPySymbol(SymbolToken):
+    __name__ = 'IonPySymbol'
+    __qualname__ = 'IonPySymbol'
+    ion_type = IonType.SYMBOL
+
+    # a good signature: IonPySymbol(ion_type, symbol_token, annotation)
+    def __new__(cls, ion_type=IonType.SYMBOL, value=None, annotations=()):
+        v = super().__new__(cls, *value)
+        v.ion_annotations = annotations
+        return v
+
+    def __copy__(self):
+        args, kwargs = self._to_constructor_args(self)
+        value = self.__class__(*args, **kwargs)
+        value.ion_type = self.ion_type
+        value.ion_annotations = self.ion_annotations
+        return value
+
+    @staticmethod
+    def _to_constructor_args(st):
+        try:
+            args = (None, (st.text, st.sid, st.location), ())
+        except AttributeError:
+            args = (None, (st, None, None), ())
+        kwargs = {}
+        return args, kwargs
+
+    @classmethod
+    def from_event(cls, ion_event):
+        if ion_event.value is not None:
+            args, kwargs = cls._to_constructor_args(ion_event.value)
+        else:
+            args, kwargs = (None, None, ()), {}
+        value = cls(*args, **kwargs)
+        value.ion_type = ion_event.ion_type
+        value.ion_annotations = ion_event.annotations
+        return value
+
+    @classmethod
+    def from_value(cls, ion_type, value, annotations=()):
+        if value is None:
+            value = IonPyNull()
+        else:
+            args, kwargs = cls._to_constructor_args(value)
+            value = cls(*args, **kwargs)
+        value.ion_type = ion_type
+        value.ion_annotations = annotations
+        return value
+
+    def to_event(self, event_type, field_name=None, in_struct=False, depth=None):
+        value = self
+        if isinstance(self, IonPyNull):
+            value = None
+
+        if in_struct:
+            if not isinstance(field_name, SymbolToken):
+                field_name = SymbolToken(field_name, 0 if field_name is None else None)
+        else:
+            field_name = None
+
+        return IonEvent(event_type, ion_type=self.ion_type, value=value, field_name=field_name,
+                        annotations=self.ion_annotations, depth=depth)
+
+
+class IonPyList(list):
+    __name__ = 'IonPyList'
+    __qualname__ = 'IonPyList'
+
+    def __init__(self, ion_type=IonType.LIST, value=None, annotations=()):
+        if value is None:
+            value = []
+        super().__init__(value)
+        self.ion_annotations = annotations
+        # it's possible to be Sexp
+        self.ion_type = ion_type
+
+    def __copy__(self):
+        args, kwargs = self._to_constructor_args(self)
+        value = self.__class__(*args, **kwargs)
+        value.ion_type = self.ion_type
+        value.ion_annotations = self.ion_annotations
+        return value
 
     @staticmethod
     def _to_constructor_args(value):
-        return (), {}
+        return (None, value,), {}
+
+    @classmethod
+    def from_event(cls, ion_event):
+        if ion_event.value is not None:
+            args, kwargs = cls._to_constructor_args(ion_event.value)
+        else:
+            args, kwargs = (None, [], ()), {}
+        value = cls(*args, **kwargs)
+        value.ion_type = ion_event.ion_type
+        value.ion_annotations = ion_event.annotations
+        return value
+
+    @classmethod
+    def from_value(cls, ion_type, value, annotations=()):
+        if value is None:
+            value = IonPyNull()
+        else:
+            args, kwargs = cls._to_constructor_args(value)
+            value = cls(*args, **kwargs)
+        value.ion_type = ion_type
+        value.ion_annotations = annotations
+        return value
+
+    def to_event(self, event_type, field_name=None, in_struct=False, depth=None):
+        value = None
+
+        if in_struct:
+            if not isinstance(field_name, SymbolToken):
+                field_name = SymbolToken(field_name, 0 if field_name is None else None)
+        else:
+            field_name = None
+
+        return IonEvent(event_type, ion_type=self.ion_type, value=value, field_name=field_name,
+                        annotations=self.ion_annotations, depth=depth)
+
+
+class IonPyDict(MutableMapping):
+    """
+    Dictionary that can hold multiple values for the same key
+
+    In order not to break existing customers, getting and inserting elements with ``[]`` keeps the same behaviour
+    as the built-in dict. If multiple elements are already mapped to the key, ``[]`  will return
+    the newest one.
+
+    To map multiple elements to a key, use the ``add_item`` operation.
+    To retrieve all the values map to a key, use ``get_all_values``.
+    """
+    __name__ = 'IonPyDict'
+    __qualname__ = 'IonPyDict'
+    ion_type = IonType.STRUCT
+
+    def __init__(self, ion_type=IonType.STRUCT, value=None, annotations=()):
+        super().__init__()
+        self.ion_annotations = annotations
+        self.__store = OrderedDict()
+        if value is not None:
+            for key, value in iter(value.items()):
+                if key in self.__store.keys():
+                    self.__store[key].append(value)
+                else:
+                    self.__store[key] = [value]
+
+    def __getitem__(self, key):
+        """
+        Return the newest value for the given key. To retrieve all the values map to the key, use ``get_all_values``.
+        """
+        return self.__store[key][len(self.__store[key]) - 1]  # Return only one in order not to break clients
+
+    def __delitem__(self, key):
+        """
+        Delete all values for the given key.
+        """
+        del self.__store[key]
+
+    def __setitem__(self, key, value):
+        """
+        Set the desired value to the given key.
+        """
+        self.__store[key] = [value]
+
+    def __len__(self):
+        return sum([len(values) for values in iter(self.__store.values())])
+
+    def __iter__(self):
+        for key in iter(self.__store.keys()):
+            yield key
+
+    def __str__(self):
+        return repr(self)
+
+    def __repr__(self):
+        return '{%s}' % ', '.join(['%r: %r' % (k, v) for k, v in self.items()])
+
+    def add_item(self, key, value):
+        """
+        Add a value for the given key. This operation appends the value to the end of the value list instead of
+        overwriting the existing value.
+        """
+        if key in self.__store:
+            self.__store[key].append(value)
+        else:
+            self.__setitem__(key, value)
+
+    def get_all_values(self, key):
+        """
+        Retrieve all the values mapped to the given key
+        """
+        return self.__store[key]
+
+    def iteritems(self):
+        """
+        Return an iterator over (key, value) tuple pairs.
+        """
+        for key in self.__store:
+            for value in self.__store[key]:
+                yield (key, value)
+
+    def items(self):
+        """
+        Return a list of the IonPyDict's (key, value) tuple pairs.
+        """
+        output = []
+        for k, v in self.iteritems():
+            output.append((k, v))
+        return output
+
+    def __copy__(self):
+        args, kwargs = self._to_constructor_args(self)
+        value = self.__class__(*args, **kwargs)
+        value.ion_type = self.ion_type
+        value.ion_annotations = self.ion_annotations
+        return value
+
+    @staticmethod
+    def _to_constructor_args(value):
+        return (None, value,), {}
+
+    @classmethod
+    def from_event(cls, ion_event):
+        if ion_event.value is not None:
+            args, kwargs = cls._to_constructor_args(ion_event.value)
+        else:
+            args, kwargs = (None, None, ()), {}
+        value = cls(*args, **kwargs)
+        value.ion_type = ion_event.ion_type
+        value.ion_annotations = ion_event.annotations
+        return value
+
+    @classmethod
+    def from_value(cls, ion_type, value, annotations=()):
+        if value is None:
+            value = IonPyNull()
+        else:
+            args, kwargs = cls._to_constructor_args(value)
+            value = cls(*args, **kwargs)
+        value.ion_type = ion_type
+        value.ion_annotations = annotations
+        return value
+
+    def to_event(self, event_type, field_name=None, in_struct=False, depth=None):
+        value = None
+
+        if in_struct:
+            if not isinstance(field_name, SymbolToken):
+                field_name = SymbolToken(field_name, 0 if field_name is None else None)
+        else:
+            field_name = None
+
+        return IonEvent(event_type, ion_type=self.ion_type, value=value, field_name=field_name,
+                        annotations=self.ion_annotations, depth=depth)
 
 
 def is_null(value):
     """A mechanism to determine if a value is ``None`` or an Ion ``null``."""
     return value is None or isinstance(value, IonPyNull)
-
-
-IonPyInt_original = _ion_type_for('IonPyInt', int, IonType.INT)
-IonPyFloat_original = _ion_type_for('IonPyFloat', float, IonType.FLOAT)
-IonPyDecimal_original = _ion_type_for('IonPyDecimal', Decimal, IonType.DECIMAL)
-IonPyText_original = _ion_type_for('IonPyText', str, IonType.STRING)
-IonPyBytes = _ion_type_for('IonPyBytes', bytes)
-IonPyList_original = _ion_type_for('IonPyList', list)
-IonPyDict_original = _ion_type_for('IonPyDict', Multimap, IonType.STRUCT)
-
-# Using the original IonPyObjects by uncomment below
-# IonPyDecimal = IonPyDecimal_original
-# IonPyNull = IonPyNull_original
-# IonPyInt = IonPyInt_original
-# IonPyFloat = IonPyFloat_original
-# IonPyText = IonPyText_original
-# IonPyList = IonPyList_original
-# IonPyDict = IonPyDict_original
-# IonPyBool = IonPyBool_original
-# IonPySymbol = IonPySymbol_original
-
-# Switch to the previous IonPyObjects by comment below lines
-IonPyDecimal = IonPyDecimal
-IonPyNull = IonPyNull
-IonPyInt = IonPyInt
-IonPyFloat = IonPyFloat
-IonPyText = IonPyText
-IonPyList = IonPyList
-IonPyDict = IonPyDict
-IonPyBool = IonPyBool
-IonPySymbol = IonPySymbol
